[Autofill] Made match field types and attributes type-safe.
This CL makes the distinction between MatchAttributes and
MatchFieldTypes in the type system much stronger:

- It promotes the two enums to enum classes.
- It replaces all bitmasks of MatchAttributes and/or MatchFieldTypes
  with DenseSet<MatchAttributes> and DenseSet<MatchFieldTypes>.
- It replaces the `int match_type`, which lumps together
  MatchAttributes and MatchFieldTypes in a single bitmask, with a
  pair of DenseSet<MatchAttributes> and DenseSet<MatchFieldTypes>.

To minimize the diff, this CL replaces `int match_type` with
`MatchParams match_type`, where MatchParams is a pair of a
DenseSet<MatchAttributes> and a DenseSet<MatchFieldTypes>. When the
AutofillPatternProvider experiment is launched, the MatchParams type
can hopefully be fully removed.

Moreover, this CL replaces the raw bitmasks in the JSON with lists
of exponents (e.g., 5 becomes [0,2]), which is less error-prone because
the exponents directly match the constants of the MatchAttributes and
MatchFieldTypes. A followup CL could add support for a string
representation.

Semantically, this CL is a no-op.

Bug: 1281115
Change-Id: I623c4e1d9b7a271e45e7cd88990d5b6cbc40204e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3347564
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#956743}