desk_templates: Move placeholder icons to the back of the container
Currently if a template has placeholder icons (grey globes), they're
left in the original order they were in before we started loading the
icons. This can create situations where the first icon in the
DesksTemplatesIconContainer is a placeholder.

This CL makes it so when a DesksTemplatesIconView loads a default icon
it will reorder itself within DesksTemplatesIconContainer to the back
of the container.

Test: manual
Bug: 1281395
Change-Id: I920da7f9d78a52f9806ae9597bdd9d61e9ab32d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3373569
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956754}