Print Preview: Fix ctrl + shift + p shortcut
e.key can be ''p', 'P', or unicode 03c0 (pi) depending on the state of
the caps lock key and whether the user is on Mac or Linux/Windows.
Update the logic accordingly.

Note: This fixes the bug but keeps the behavior different for other
keyboard layouts as compared to M97 - the current layout's p key will
be used instead of the physical p key. This will match the print
preview keyboard shortcut for the system print dialog to the shortcut
used in the general browser for the same action.

Fixed: 1284598
Change-Id: I733cea8117a716659b3da1b1d1e7a293df40edad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3368902
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956798}