StorageBuckets: Deletion callback to return after DB bucket deletion
This change updates Bucket deletion logic to return QuotaStatus
after deletion from QuotaDatabase. StorageKeyDeletion did not wait for
deletion to complete because entries for storage_key in the
QuotaDatabase were created and removed silently. But this should be
different for buckets because bucket data will continue to appear for
the Buckets API as long as it still exists in the QuotaDatabase.
Therefore deletion success should be returned after we know deletion
from the QuotaDatabase was successful.

Change-Id: Ic229678fa38125945883a446bccc095f76820faa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3272085
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956802}