[Launcher pop-in] Track the burn-in iteration for categories.
We have a pre-existing burnin_iteration_counter_ which is applied to
individual results. In this CL, we expand the concept to cover the
(simpler) case of categories. Record the burn-in iteration number for a
category the first time it is seen during a search.

This is applicable to query search, and not to zero-state search.

A follow-up CL will use the burn-in iteration number as part of
category sorting. I've separated the record-keeping and the sorting into
separate CLs because I find it makes the test updates conceptually
easier to follow.

Test: unit tests
Bug: 1279686
Change-Id: I70048456d5bc7127b7afe03f357b15fa14278d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372092
Reviewed-by: Tony Yeoman <tby@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956947}