MacPWAs: Fix keyboard navigation of 3-dot menu in PWAs
Keyboard navigation of the 3-dot menu is not done by NSEvents arriving
at the NSView or NSWindow of the menu's Widget. Rather, a
MenuPreTargetHandlerMac is created, which calls +[NSEvent
addLocalMonitorForEventsMatchingMask:handler:], which will intercept
NSEvents before they are dispatched.

This doesn't work for PWAs, because the NSEvent monitor is created
in the browser process, not in the PWA process.

Add a function NativeWidgetMacNSWindowHost::AddEventMonitor, which
will create the monitor in the correct process.

This is done in a generic way because, by inspection, there are
several other places where addLocalMonitorForEventsMatchingMask is
called in the wrong place, and the more generic framework will be
required.

Bug: 1254454
Change-Id: Ib1f5415155477ba6557aa651c93153b1d9a4c527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3348103
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956952}