Fix DCHECK in search_result_ranker.cc
In the transition between query search and zero-state, there are cases
where the results list is improperly cleared leading to query results
remaining in the results list. This doesn't affect the UI because these
results are filtered at a later step of the pipeline, but does
sometimes trigger this DCHECK. This CL simply removes it for now and
doesn't do any extra zero-state scoring logic for non-zs results.

crbug.com/1284507 is tracking more significant changes to how the
results list is handled, which should fix the root cause of the issue.

Bug: 1282329
Change-Id: I26e2dc412b015552c3be79153948ac0871a422fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3376025
Reviewed-by: Rachel Wong <wrong@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956963}