Don't get stuck with legacy layout after printing.
We need to fall back to legacy layout when printing (as long as
LayoutNGPrinting isn't enabled), but this shouldn't affect legacy
fallback for the non-printed document.

Just reset the legacy fallback flag on the root element when toggling
print mode, and hope that style is recalculated for all elements (in
case there's anything in there that requires us to do legacy fallback
for the entire document, even if we aren't printing).

Bug: 1285208
Change-Id: I68ce9886ec4ccd0fcf37765fa1f98b29ca710d72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372961
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956970}