[Autofill Assistant] Pass suppliers instead of ActivityTabProvider
We are in the process of moving all Autofill Assistant code to //components. For this we need to remove all dependencies on Chrome code.

In this cl we pass in suppliers for the Activity and the WebContents
instead of Chrome's ActivityTabProvider.

Bug: b/173103628
Change-Id: I50454ae8ba5794a77a7f6d5c227f7086f8520ddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3367626
Auto-Submit: Florian Gauger <fga@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Florian Gauger <fga@google.com>
Cr-Commit-Position: refs/heads/main@{#956995}