[sync] Inline test-only RetryVerifier into ExponentialBackoffChecker
No behavioral changes: this patch moves the code in RetryVerifier as is
into ExponentialBackoffChecker. The split between the two classes is
artificial and not worth maintaining, sine it harms the checker's
ability to report more detailed failure messages that would be useful
when debugging test flakiness.

Change-Id: I147eb02a7e38a1989d2a23ea5deb368780859a3c
Bug: 1170609
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3376344
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957007}