Reduce the use of the “color” attribute in perf_tests.css.
The color attribute is used several places as a neutral attribute in
e.g. tests related to selector performance. But now that color will
get special properties (it will be one of very few independently
inherited attributes), it is a special case and no longer well-suited
for benchmarking. We'll use the cursor attribute instead to fill
the same role (it's inherited, but not marked as independent).

Some benchmark changes will inevitably occur from this change,
as we're modifying what the benchmark is measuring.

Bug: 1284225
Change-Id: Ic1c0140b2eaaeb6d31d4c41e6d439bf4ab40e923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3377062
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957052}