[Lacros] Re-enable InitialBrowserAction::kDoNotOpenWindow
In order to share code with other platforms, the parameter is converted
to the switches::kNoStartupWindow command line switch.

Note that the logic could not just be simply kept in
StartupBrowserCreator::ShouldLoadProfileWithoutWindow():
unlike the command line parameter, the kDoNotOpenWindow action remains
set for the whole execution of Chrome. However, it is intended to
only affect startup.

An alternative would have been to:
- plumb another flag to check whether this is startup, e.g.
  IsProcessStartup
- convert all callsites to pass this flag. In particular some callsites
  seem to pass the wrong value of IsProcessStartup, so they would need
  to be updated (e.g. in profile_window.cc).

Fixed: 1278549
Change-Id: I88e5046497b750e4f0bc3a1afebcbb41d2dcbd8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3358211
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957068}