clang scripts: Make DownloadUrl() understand gzip encoding
Without `Accept-Encoding: gzip`, we don't actually get the smaller data.
Instead, it's just-in-time decompressed on the server now.
Now, we just download 1.6MB instead of 11MB and decompress on the client
instead. (I verified that after decompression, we produce the same bytes
as both `curl` and `curl --compressed`, which do server-side and
client-side decompression too, respectively.)

This has the added benefit that we now get a `Content-Length` header
in the response again, which happens to unbreak the ToT bots.

Bug: 1267363
Change-Id: If7b49952d164fc2df963edf50a9f00b882772563
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3377922
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957081}