[ios] Fix crash when stopping a download
The DownloadNativeTaskBridge owns and observe (using KVO) the
WKDownload object. The WKDownload causes KVO events to fire
when it is deallocated which retain DownloadNativeTaskBridge
during its -dealloc.

This causes a crash as a object that is deallocated must never
be retained (best case, our zombie detector will warn about
this, worse case it will lead to use-after-free).

To fix this, stop the observation when the download is over,
cancelled or when the DownloadNativeTaskBridge object is
deallocated (i.e. before WKDownload is destroyed).

Bug: 1282244
Change-Id: I4af7683940f05339af8b3451c30e624fd90dd492
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3377184
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957095}