Android: Extend the fullscreen toast duration when defocused
If the fullscreen toast message becomes invisible due to other window
stealing the focus, pause the timer of the toast and resume it
when the focus is regained. In this way we can guarantee that the toast
will be visible to users for the fixed amount of time.

Bug: 1270052
Change-Id: I4f80e160cb8b230de48473afbaf782443b6ff27d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372521
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957110}