webauthn: accept a new caBLEv2 AEAD construction.
We would like to make the caBLEv2 messages look more like standard Noise
messages. Currently they use a little-endian nonce and have an
additional data of [2], which is a carry-over from caBLEv1.

As a first step to moving to big-endian nonces and no additional data
(which is what standard Noise does), this change recognises both forms
in the first message and switches accordingly. Once this has gone
everywhere, in some number of months, we can switch the default form
that is sent to complete the transition.

BUG=1002262

Change-Id: I3686e514b3e782366ed6a427b478989f9be7c723
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3370809
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957170}