[profiles] Convert SwitchToProfile to OnceCallback
This CL is a simple cleanup with no behavior changes.
It converts from RepeatingCallback to OnceCallback and removes the
CreateStatus argument which was not useful.

The main changes are in profile_window.[h/cc]
The other files are only updated with the new API.

Change-Id: Iec0993db4e645d497de647aa4fdec9e42473f52b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372763
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957173}