extensions: add WebAuthn registration proxying to webAuthenticationProxy
This adds an onCreateRequest event and completeCreateRequest method to
the webAuthenticationProxy extensions API.

While running with an attached webAuthenticationProxy extension,
WebAuthn navigator.credentials.create() calls raise a onCreateRequest
event to that extension. The event contains a JSON-serialized
representation of the PublicKeyCredentialCreationOptions passed to the
WebAuthn API call.

The extension calls completeCreateRequest() to inject a JSON-serialized
representation of the PublicKeyCredential, which Chrome returns to the
calling site.

For JSON serialization, all ArrayBuffer-valued fields are replaced with
base64url-encoded DOMStrings. We plan to provide compatible counterpart
parsing/serilization methods as part of WebAuthn.

Bug: 1231802
Change-Id: If727a98154bf3d72bc77d4ef1edcc834a47709f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3311658
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Charles Reis <creis@chromium.org>
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Auto-Submit: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/main@{#957177}