Dictation: Record when macros succeed and fail
This change adds behavior to the Dictation extension that tracks
whenever a macro succeeds or fails. This change adds a test for the
success metric, but not for the failure metric. This is because our
parsing and context checking is still too naive to ever result in a
failed macro. A test for the macro failure metric will be added when
we have more complex parsing (using pumpkin).

AX-Relnotes: N/A
Bug: 1247299
Change-Id: Ibe68c44ffae409942d1f4eecd2d8b135196c7259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3367236
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957180}