Fix scale_factor getting reset
When Windows is set to a scale factor other than 1.0, closing or
opening laptop with an external mirrored monitor can cause scale factor
to be set to the default. This CL makes the code not overwrite the scale
factor when GetDisplayNearestWindow() returns the default monitor in
WindowTreeHost::OnHostResizedInPixels().

The underlying issue is that opening and closing the laptop changes
the device returned by MonitorInfoFromWindow, which means
ScreenWin::GetScreenWinDisplay returns the default display.

Bug: 1285476
Change-Id: I5079b0092788dd7bec68fffd51350877414c4897
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3371754
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: David Bienvenu <davidbienvenu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957188}