PageInfo: Fix flakiness by creating histogram snapshots
Batched tests are failing because forgetHistogramForTesting() prevents
histograms to be recorded after it was called. To fix this, a
HistogramTesterRule class is added that creates snapshots similar to the
C++ HistogramTester class. This replaces forgetHistogramForTesting() as
it allows for easier and more reliable resets of Histogram state.

Bug: 1211884, 1282505
Change-Id: I3788c90ef426fd122950e05813947a004bd46375
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3366427
Reviewed-by: Donn Denman <donnd@chromium.org>
Reviewed-by: Ehimare Okoyomon <eokoyomon@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957246}