Switch ShouldRecordHistogram() to a reader lock.
Similar to FindHistogram(), this doesn't actually modify
state and can be a reader lock. It previously had a call to
EnsureGlobalRecorderWhileLocked(), but this is unnecessary -
even if it ended up creating |top_|, that object would not
have a |record_checker_| set.

I noticed this showed up as causing contention on some
reports of crbug.com/1284513

Bug: 1286050, 1123627
Change-Id: Ida085ecf45065121f5217bbf3b7263a4fb4849cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3378925
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957262}