Dictation: Set z order level to ensure bubble UI is user-visible
This change fixes an issue where the Dictation bubble UI was
occasionally hidden behind the Chrome and Settings windows, and thus
not visible to the user. This change fixes that issue by setting the
z order level to ensure the UI gets displayed as a floating UI element.

AX-Relnotes: N/A
Bug: 1252037
Change-Id: I4e8a03c7b8515117287f90f462fe6d04c2e277cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3372663
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957267}