Mac: Ignore ENOTCONN from guarded_close_np()
The Mac OS kernel has a bug where guarded_close_np() may return an ENOTCONN error. This is impossible to handle correctly because we don't know if the socket is still open or not. Previously we crashed when this happened, but the crash frequency became too high. Potentially leak the file descriptor instead.

BUG=1151048

Change-Id: I7f098005be01a93811077c31e5f3d59e43341360
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3370122
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957350}