desk_templates: Change layouting for name view and admin indicator.
Use a box layout view to parent them both.
Add a GetAvailableWidth function since GetAvailable size does not
work with BoxLayout.
Remove minimum length for template name.
Fixes issue that tabbing doesn't always save the name and may lead to
an eventually crash. This is due to a bug where the grid was not
being recreated (though we want to eventually not recreate) and using
the old template name.

Test: manual
Bug: 1264174, 1281371, 1285113
Change-Id: Ic7ba2a69c558e8fefad728e5442ce7ef9d4e9c1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3353864
Reviewed-by: Jeremy Chinsen <chinsenj@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957363}