[LayoutNG] Substructs for break token data
We create a special break token data structure for grid and flexbox.
However, we will never need both of these data structures at the same
time.

This CL fixes this by creating a new NGBlockBreakTokenData struct that
is stored on the NGBlockBreakToken. NGFlexBreakTokenData and
NGGridBreakTokenData are now substructs of NGBlockBreakTokenData,
allowing us to have one data struct per break token.

The reason that I chose to subclass a member variable of the break
token rather than the break token itself is that NGBlockBreakToken
holds a flexible array, making it tricky to add any subclass. I'd be
open to testing out other ideas, as well.

Bug: 660611,614667
Change-Id: I82355239b935646f3ac742e9295a7271339f3321
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3366808
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#957373}