cros: Fix the crash of move to desk menu.
The crash is caused by the DCHECK in MapCommandIdToDeskIndex. Because
the title command id -2 is less than move to first desk command id
40801. This happens when calling
MoveToDesksMenuDelegate::IsCommandEnabled with title command id.

To solve this, the cl checks if the command is a "move to desk"
command before calling MoveToDesksMenuDelegate::IsCommandEnabled.

Bug: 1284702
Change-Id: Ib10c8fd239d3ff957aab9f8811acee85c9641956
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3378723
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957382}