[Files app] Open new tab in lacros when lacros is primary
Files app has an OpenNewTab method that it used to open a new tab in
the browser and navigate to a given URL.

New tabs are currently always opened in the ash-chrome browser. Change
the behavior to open a lacros-chrome tab when the `lacros-primary` flag
is enabled.

Tested opening PDF and HTML files when #lacros-primary is
enabled/disabled - a tab is opened in the right browser.

Test: OpenNewTab is used when Files app uses Chrome as file handler.
Bug: chromium:1260237
Change-Id: Ib17adc8d561fd3dc3b46d8af3a1a29bfabcbf9a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3283907
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#957397}